[Bug Fix] Fix losing files
1. Recommit and merge losing files on push-to-hub feature
2. Fix a failure to build a int tensor with float values
Link: https://code.alibaba-inc.com/Ali-MaaS/MaaS-lib/codereview/12308629

+++ modelscope/trainers/hooks/logger/text_logger_hook.py
@@ -75,7 +75,7 @@ class TextLoggerHook(LoggerHook):
     def _get_max_memory(self, trainer):
         device = torch.cuda.current_device()
         mem = torch.cuda.max_memory_allocated(device=device)
-        mem_mb = torch.tensor([mem / (1024 * 1024)],
+        mem_mb = torch.tensor([int(mem / (1024 * 1024))],
                               dtype=torch.int,
                               device=device)
         if trainer._dist:
